+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Changelog
 - Keyboard support, ESC key toggle overlay
+- Add default image path (./overlay/) for projects without WordPress
 
 ## [0.1.0] - 2018-07-14
 ### Added